npm run dev works but build throws invalid default export on (posts)/layout.tsx
