+++ app/(posts)/__example/page.mdx
@@ -1,0 +1,12 @@
+export const postDetails = {
+  title: "Title here",
+  standfirst: "Standfirst here.",
+  author: "Baird Langenbrunner",
+  publishDate: "2025-09-24",
+  tags: ['tag1', 'tag2']
+}
+
+import PostHeading from '@/app/components/postheading.tsx'
+import Triangles from '@/app/components/triangles.tsx'
+
+<PostHeading metadata={postDetails} />
