+++ app/(posts)/first-post/page.mdx
@@ -58,11 +58,13 @@ IPCC reports. [AR5](https://www.ipcc.ch/assessment-report/ar5/), for example, wa
     Policymakers](https://www.ipcc.ch/report/ar5/wg1/summary-for-policymakers/),
     showing low- and high-emissions climate trajectories for temperature,
     precipitation, sea ice extent, and ocean surface pH. This was one of the
-    first times the Robinson projection was used in a report like this.
+    first times the Robinson projection and smarter graphics were used in a report like this.
   </figcaption>
 </figure>
 
-Those graphics still have a long way to go to be accessible to the general public, I think, and to be readable on cell phones... but that's not ultimately why they're being created.
+Little by little, map projections and color schemes like these have trickled into academic publications and conference presentations. It's been a great thing to witness, and I remember holding high standards for figures when I worked as an academic editor myself, regularly asking authors to modify things to make them more accessible, and to read up on why this is important.
+
+These IPCC-style graphics still have a long way to go to be accessible to the general public, I think, and to be readable on cell phones... but that's not ultimately why they're being created.
 
 ### But I'd like to be
