--- app/(posts)/second-post/page.mdx
@@ -1,15 +1,0 @@
-export const postDetails = {
-  title: "This is my second post",
-  standfirst: "Second post standfirst here.",
-  author: "Baird Langenbrunner",
-  publishDate: "2024-06-17",
-  tags: ['tag1', 'tag2']
-}
-
-import PostHeading from '@/app/components/postheading.tsx'
-
-<PostHeading metadata={postDetails} />
-
-### Second post
-
-This is an example of a second post.
