+++ package.json
@@ -17,7 +17,8 @@
     "next": "14.2.4",
     "next-mdx-remote": "^5.0.0",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-tweet": "^3.2.1"
   },
   "devDependencies": {
     "@types/mdx": "^2.0.13",
